Split test_table.py into multiple files
