+++ python/src/hash.py
@@ -2,6 +2,8 @@ from src.dptables import *
 
 
 def hash_quinary(hand, lenbit, k):
+  if isinstance(hand, int):
+    hand = list(map(int, reversed(str(hand))))
   hash_ = 0
   for i in range(lenbit):
     hash_ += DP[hand[i]][lenbit-i-1][k]
